+++ src/lib.rs
@@ -15,7 +15,7 @@
 //! To use this, add it as a dependency to your Cargo.toml:
 //! ```toml
 //! [dependencies]
-//! noise_fn = "^0.1.1"
+//! noise_fn = "^0.1.2"
 //! ```
 
 #![no_std]
@@ -24,12 +24,13 @@
 #![feature(associated_type_defaults)]
 #![feature(external_doc)]
 
-#![doc(html_root_url = "https://docs.rs/noise_fn/0.1.1")]
+#![doc(html_root_url = "https://docs.rs/noise_fn/0.1.2")]
 
 pub mod noise;
 pub mod helpers;
 pub mod perm_table;
 mod white;
+mod hash;
 mod simplex;
 mod scale;
 mod add;
@@ -37,6 +38,7 @@ mod constant;
 mod gradient;
 mod sum;
 mod octaves;
+mod to_float;
 
 pub use {
 	noise::{
@@ -46,6 +48,7 @@ pub use {
 	},
 	helpers::Config,
 	white::WhiteNoise,
+	hash::HashNoise,
 	simplex::Simplex,
 	scale::ScaleNoise,
 	add::AddNoise,
@@ -53,6 +56,7 @@ pub use {
 	gradient::Gradient,
 	sum::SumNoise,
 	octaves::Octaves,
+	to_float::ToFloat,
 };
 
 // Include the readme and changelog as hidden documentation so they're tested by cargo test